+++ lib/active_attr/typecasting/unknown_typecaster_error.rb
@@ -1,0 +1,13 @@
+require "active_attr/error"
+
+module ActiveAttr
+  module Typecasting
+    # This exception is raised if attempting to cast to an unknown type when
+    # using {Typecasting}
+    #
+    # @since 0.6.0
+    class UnknownTypecasterError < TypeError
+      include Error
+    end
+  end
+end